+++ src/shared/ui/form_inputs/field_wrappers/pilled_wrapper/pilled_wrapper.css
@@ -1,17 +1,7 @@
-.pilled-input {
+.pilled-wrapper {
   display: flex;
   flex-direction: column;
   gap: var(--size-1);
-
-  &__feedback {
-    &_danger {
-      color: var(--clr-danger-5);
-    }
-  }
-
-  &__required-sign {
-    color: var(--clr-danger-5);
-  }
 
   &__autosuggestions {
     display: flex;
