+++ src/shared/ui/form_inputs/field_wrappers/pilled_wrapper/pilled_wrapper.tsx
@@ -1,4 +1,4 @@
-import './pilled_input.css';
+import './pilled_wrapper.css';
 import { For, type ParentProps, Show } from 'solid-js';
 
 export interface PilledWrapperProps extends ParentProps {
@@ -10,12 +10,12 @@ export function PilledWrapper(props: PilledWrapperProps) {
   return (
     <div class="pilled-wrapper">
       <Show when={props.pills.length != 0}>
-        <div class="pilled-input__autosuggestions">
+        <div class="pilled-wrapper__autosuggestions">
           <For each={props.pills}>
             {(pill) => {
               return (
                 <span
-                  class="pilled-input__pill pilled-input__pill_brand"
+                  class="pilled-wrapper__pill pilled-wrapper__pill_brand"
                   onClick={() => props.onPillClick?.(pill)}
                 >
                   {pill}
