+++ src/shared/ui/form_inputs/full_inputs/README.md
@@ -1,3 +1,8 @@
 # full_inputs
 
 Здесь лежат непосредственно элементы, являющиеся кусками форм на верхнем уровне.
+**Не кладите сюда ничего, если это не используется повсеместно. Рекомендуется собирать форму из отдельных кусочков из fields и field_wrappers.**
+
+## Input
+
+Элемент Input оставлен для обратной совместимости с login и register формами. Не используйте в новых формах
