+++ src/shared/ui/form_inputs/index.ts
@@ -1,3 +1,6 @@
-export { FieldState } from './fields/field_state';
-export { Input } from './full_inputs/input/input';
+export { FieldState, HiddenInputField, InputField, SelectField, TextareaField } from './fields';
+export { Input } from './full_inputs/input';
 export { transformFieldState } from './tanstack_adapter';
+
+export { LabelsWrapper } from './field_wrappers/labels_wrapper/labels_wrapper';
+export { PilledWrapper } from './field_wrappers/pilled_wrapper/pilled_wrapper';
