+++ src/shared/ui/index.ts
@@ -4,15 +4,9 @@ export { List } from './list/list';
 export { P } from './p/p';
 export { Button, ButtonStyle } from './button/button';
 export { Navbar } from './navbar/navbar';
-export { Input, FieldState, transformFieldState } from './form_inputs';
-// export { InputField, InputState } from './input_field/input_field';
-// export { HiddenInputField } from './hidden_input_field/hidden_input_field';
-// export { TextareaField } from './textarea_field/textarea_field';
-// export { Textarea } from './textarea/textarea';
+export * from './form_inputs';
 export { Tabbar, type TabbarOption } from './tabbar/tabbar';
 export { AuthGuard } from './auth_guard/auth_guard';
-// export { SelectField } from './select_field/select_field';
-// export { PilledInput } from './pilled_input/pilled_input.tsx';
 
 export { CalendarIcon } from './icons/calendar';
 export { UserIcon } from './icons/user';
